requirements.txt
librosa
numpy~=1.24.1
PySoundFile
soundfile~=0.12.1
torch~=2.1.2+cu121
torchaudio~=2.1.2+cu121
torchviz
tqdm~=4.66.1
ffmpeg-python
scikit-learn~=1.3.2